# backend/school-management/public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management API",
        "description": "Documentación de la API para el sistema de gestión escolar",
        "contact": {
            "name": "Equipo de desarrollo CBTa71",
            "email": "soporte@cbta71.edu.mx"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000/api",
            "description": "Servidor local"
        },
        {
            "url": "https://nginx-production-728f.up.railway.app/api",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
...
        "/api/v1/find/concept/{id}": {
            "get": {
                "tags": [
                    "FindEntity"
                ],
                "summary": "Buscar concepto de pago por ID",
                "description": "Obtiene la información de un concepto de pago específico mediante su identificador.",
                "operationId": "b916b8cfed1a9f3c7ffdeaa1eac6a8ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del concepto a buscar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concepto encontrado correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "concept": {
                                                    "$ref": "#/components/schemas/DomainPaymentConcept"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Concepto encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Concepto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "El concepto solicitado no existe."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error inesperado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/find/payment/{id}": {
            "get": {
                "tags": [
                    "FindEntity"
                ],
                "summary": "Buscar pago por ID",
                "description": "Obtiene la información detallada de un pago específico mediante su identificador.",
                "operationId": "e17c318789b79b846f82d25d0d75c349",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pago a buscar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pago encontrado correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "payment": {
                                                    "$ref": "#/components/schemas/DomainPayment"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pago encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pago no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "El pago solicitado no existe."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error inesperado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/find/user": {
            "get": {
                "tags": [
                    "FindEntity"
                ],
                "summary": "Obtener usuario autenticado",
                "description": "Devuelve la información del usuario autenticado en el sistema.",
                "operationId": "ebfc31a0a1f630400906aec03096a4c6",
                "parameters": [
                    {
                        "name": "forceRefresh",
                        "in": "query",
                        "description": "Forzar actualización del caché (true o false).",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario autenticado encontrado correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/DomainUser"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontró el usuario autenticado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error inesperado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DomainPayment": {
                "description": "Representa un pago realizado por un usuario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 123
                    },
                    "payment_concept_id": {
                        "type": "integer",
                        "example": 45,
                        "nullable": true
                    },
                    "payment_method_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "stripe_payment_method_id": {
                        "type": "string",
                        "example": "pm_1Hh1Xx2eZvKYlo2Cj1234567",
                        "nullable": true
                    },
                    "concept_name": {
                        "type": "string",
                        "example": "Pago de inscripción",
                        "nullable": true
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 1500,
                        "nullable": true
                    },
                    "payment_method_details": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Tarjeta de crédito",
                            "Banco XYZ"
                        ],
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "pendiente"
                    },
                    "payment_intent_id": {
                        "type": "string",
                        "example": "pi_1Hh1Xx2eZvKYlo2Cd1234567",
                        "nullable": true
                    },
                    "url": {
                        "type": "string",
                        "example": "https://checkout.stripe.com/pay/cs_test_a1b2c3d4",
                        "nullable": true
                    },
                    "stripe_session_id": {
                        "type": "string",
                        "example": "cs_test_a1b2c3d4",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "DomainPaymentConcept": {
                "description": "Representa un concepto de pago",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "concept_name": {
                        "type": "string",
                        "example": "Pago de inscripción"
                    },
                    "description": {
                        "type": "string",
                        "example": "Pago correspondiente al semestre 2025A",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "activo"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-09-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31",
                        "nullable": true
                    },
                    "amount": {
                        "type": "string",
                        "example": "1500.00"
                    },
                    "applies_to": {
                        "type": "string",
                        "example": "todos"
                    },
                    "is_global": {
                        "type": "boolean",
                        "example": true
                    },
                    "userIds": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    },
                    "careerIds": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    },
                    "semesters": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
            "DomainStudentDetail": {
                "description": "Detalles del estudiante",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 123
                    },
                    "career_id": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "n_control": {
                        "type": "integer",
                        "example": 20201234,
                        "nullable": true
                    },
                    "semestre": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "group": {
                        "type": "string",
                        "example": "A",
                        "nullable": true
                    },
                    "workshop": {
                        "type": "string",
                        "example": "Taller de programación",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "DomainUser": {
                "description": "Representa un usuario del sistema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "hashed_password"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+5215512345678"
                    },
                    "birthdate": {
                        "type": "string",
                        "format": "date",
                        "example": "1995-06-15",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "example": "male",
                        "nullable": true
                    },
                    "curp": {
                        "type": "string",
                        "example": "PEMJ950615HDFRZN09"
                    },
                    "address": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Calle Falsa 123",
                            "Colonia Centro"
                        ],
                        "nullable": true
                    },
                    "stripe_customer_id": {
                        "type": "string",
                        "example": "cus_ABC123XYZ",
                        "nullable": true
                    },
                    "blood_type": {
                        "type": "string",
                        "example": "O+",
                        "nullable": true
                    },
                    "registration_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T12:34:56Z",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "activo",
                        "nullable": true
                    },
                    "studentDetail": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/DomainStudentDetail"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}